fix telemetry & cleanup
